backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
agno>=1.7.0
sse-starlette>=2.3.0
aiosqlite>=0.20.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.11.0
google-generativeai>=0.8.0
httpx>=0.27.0
python-multipart>=0.0.18
aiofiles>=23.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
